customUserDetailsService class finish -> remove localhost first login page 24/01/09

+++ src/main/java/joo/community/dto/user/UserDto.java
@@ -8,14 +8,14 @@ import lombok.Data;
 @Data
 public class UserDto {
 
-    private int id;
+    private Long id;
     private String username;
     private String name;
     private String nickname;
 
     // 'Cannot be accessed from outside package' 에러 발생으로 인한, 생성자 추가.
     // 명시적으로 public 생성자 추가 후 디버깅 완료
-    public UserDto(int id, String username, String name, String nickname) {
+    public UserDto(Long id, String username, String name, String nickname) {
         this.id = id;
         this.username = username;
         this.name = name;
